✨ Compatibility | v1.7.0 ✨
This update includes an compatibility remake for better connection with the latest NextLanguage v1.6 & the upcoming v1.7

This Commit was made for an unofficial version of NextLibrary.

------------------------------------

With the full files displayed on
the source code. With no hidden
files on the downloaded release.

-----------------------------------

This is an better maintained
Repositories of NextLibrary.

Forked from the official
Repositories from Nxium's
Official NextLibrary's Repo